refactor angular adapter code
refactor to move helpers to utils.ts, and add interface.ts
