+++ packages/@apphosting/adapter-angular/src/interface.ts
@@ -1,0 +1,9 @@
+export interface OutputPathOptions {
+    bundleYamlPath: string;
+    outputDirectory: string;
+    baseDirectory: string;
+    outputBaseDirectory: string;
+    serverFilePath: string;
+    browserDirectory: string;
+  }
+
