+++ packages/@apphosting/adapter-angular/src/utils.ts
@@ -1,6 +1,10 @@
 import fsExtra from "fs-extra";
 import { fileURLToPath } from "url";
 import type { ApplicationBuilderOptions } from "@angular-devkit/build-angular";
+import { spawn } from "child_process";
+import { resolve, normalize, relative } from "path";
+import { stringify as yamlStringify } from "yaml";
+import { OutputPathOptions } from "./interface.js";
 
 // fs-extra is CJS, readJson can't be imported using shorthand
 export const { writeFile, move, readJson } = fsExtra;
@@ -58,6 +62,71 @@ export async function loadConfig(cwd: string): Promise<ApplicationBuilderOptions
 
   return applicationBuilderOptions;
 }
+
+// populate file/directory paths we need inside app hosting output directory
+export function populateOutputBundleOptions(config: ApplicationBuilderOptions): OutputPathOptions {
+  const outputPath = config.outputPath;
+  // normalized output path structure
+  const normalizedOutputPath = {
+    browser: "browser",
+    server: "server",
+    media: "media",
+    ...(typeof outputPath === "string" ? undefined : outputPath),
+    base: normalize(resolve(typeof outputPath === "string" ? outputPath : outputPath.base)),
+  };
+  const outputBundleDir = resolve(".apphosting");
+  
+  return {
+    bundleYamlPath: resolve(outputBundleDir, "bundle.yaml"),
+    outputDirectory: outputBundleDir,
+    baseDirectory: resolve(normalizedOutputPath.base),
+    outputBaseDirectory: resolve(outputBundleDir, "dist"),
+    serverFilePath: resolve(outputBundleDir, "dist", normalizedOutputPath.server, "server.mjs"),
+    browserDirectory: resolve(outputBundleDir, "dist", normalizedOutputPath.browser),
+  };
+}
+
+// Run build command
+export const build = (cwd = process.cwd()) =>
+  new Promise<void>((resolve, reject) => {
+    const process = spawn("npm", ["run", "build"], { cwd, shell: true, stdio: "inherit" });
+    process.on("exit", (code) => {
+      if (code === 0) return resolve();
+      reject();
+    });
+  });
+
+// move the base output directory, which contains the server and browser bundle directory, and prerendered routes
+// as well as generating bundle.yaml
+export async function generateOutputDirectory(
+  cwd: string,
+  outputPathOptions: OutputPathOptions,
+): Promise<void> {
+
+  await Promise.all([
+    move(outputPathOptions.baseDirectory, outputPathOptions.outputBaseDirectory, { overwrite: true }),
+    generateBundleYaml(outputPathOptions, cwd),
+  ]);
+}
+
+// generate bundle.yaml
+export async function generateBundleYaml(
+  outputPathOptions: OutputPathOptions,
+  cwd: string,
+): Promise<void> {
+  await writeFile(
+    outputPathOptions.bundleYamlPath,
+    yamlStringify({
+      headers: [],
+      redirects: [],
+      rewrites: [],
+      runCommand: `node ${normalize(relative(cwd, outputPathOptions.serverFilePath))}`,
+      neededDirs: [normalize(relative(cwd, outputPathOptions.outputDirectory))],
+      staticAssets: [normalize(relative(cwd, outputPathOptions.browserDirectory))],
+    }),
+  );
+}
+
 
 export const isMain = (meta: ImportMeta) => {
   if (!meta) return false;
